[Stable] Completely rewritten! MVVM!
As it turned out the MVVM pattern is nice for switching views, and with
what i have in mind switching views for settings would be a pain without
it. I used MVVM Light Toolkit and some Nugget Packages for the Folder
Dialog. I made some changes to Bindings as well (MVVM typical) with
INotifyPropertyChanged so we don't need to refresh the listview
ourselves.

TL;DR Completely rewritten to MVVM

- Add
Nugget API PAckages
documentation of Nugget Packages used

- Fix
Random Exception when calling Viewmodel
_empList was not accessible

- Changed
Data bindings for listview