FilterList, Templates, Output.tex, Regex
The only good way to find and replace things in a string is RegEx, so i
went with that. Inside the templates there are placeholder which will
get replaced with the desired string. Example for a placeholder
$Language$. I also added a filterlist with some basic endings in it,
which should cover most files.

- Add
Templates for Code, Images, PDF
FilterList with some extensions
RegEx for Placeholders in Templates
Basic Exception handling for files

- Change
RegEx works with the $$ placeholder

- Fix
Exceptions would show up due to incorrect file handling

- Bug
Dropdownbutton list is not selectable