Major Update
This update revolves about multiple instance handling, switching views,
views, the whitelist, adding stuff to the whitelist and code cleanup.

If you have another instance of the programm open and you try to start
it, you will be greeted with a messagebox.

Since the auto update on propertychanged event didndt work correctly
with the grid, I had to set the itemsource to null and to the list
again.

The SwitchViewModel can now feature several views depending on the
ViewModel given to it via DataContext, but for this to work you have to
make a DataTemplate with the ViewModel name and the View in it.

I added a new View AddextensionView. This view allows to add your own
extensions. with the maskedTextBox your input will be filtered. The
added extension is not added to the whitelist file yet, but it will be
added to the whiteList in the main programm. For now you can just add
extensions to the WhiteList file.

WhiteList now implements IComparable, which will look at the Extension.

Moved several classes to the Model Folder, as they weren't helper
classes but Models.

The MainView now contains a whitelist grid and a combobox for the
languages / extensions. If you select a Language, the grid will only
show those extension in the same language category.

- Add
Views: AddExtensionView, SettingsView
ViewModels: AddExtension, SettingsViewModel
WhiteList class
Multiple Instance Handling
Extended WPF Toolkit for maskedTxtBox

- Change
SwitchViewWindow can now switch between Views
SwitchViewWindow now starts in the middle of the owner (Main)
Databindings for the Grids

- Fix
SwitchViewWindow couldn't be closed
Exceptions when trying to save tex

Andi