Message Update, Code Documentation, & more
I added "notifications" so the user can see that sth actually happens,
if he is pressing buttons. Although it is not possible to update the
databinding for the view in the view.cs, i made a workaround with an
ICommand which then raises the PropertyChangedEvent while still having
the click event.

To make it easier for contributors, comments were added.

- Add
Notifications
Global Exception Handling
Crash log + formatting

- Changed
Anchors for elements in the UI so resizing the UI wont cause problems
anymore

- Fix
Notifications didnt show up when propertychanged event was raised
Resizing the UI would cause overlay problems
Crash Logs were overwritten everytime an exception occurred