MenuItem, Image, Some fixes
Added new MenuItems for the Grids, if you right click on the
WhiteListGrid or the MainGrid you will get a clear or remove item
option. Also added a check for the WhiteList.txt if there is no txt aa
new one with standard variables will be generated.

- Add
MenuItems (Clear, Remove Item)
MenuItem Icon for exit
WhiteList.txt check

- Change
Exit actually closes the application now.

- Bug (LaTex
Since LaTex does not support whitespaces in paths you will get an error
if a whitespace is in the paths in output.tex. I will add a warning for
the user so he knows that a path with whitespaces is in the list.
Apparently there is no workaround for this problem in LaTex.