Bug fixes, settings, replace text, contex menu items, file check
Added a check for all the needed files, if none are existing all of them
will be generated with standard values. Finaly added a use for the
settings view - although there are only two options for now,
SaveWhiteList and StatusBar texts. The TexEditorView is now autosizing
on the owner window. The exit menuitem works, although i have to close
it the bad way, somehow there is still a thread alive and i cant figure
out which one yet so i had to use Environment.Exit(1). Furthermore i
added a picture to the item. Fixed the two "all" options in language box
bug if nothing was in whitelist.txt

- Add
MenuItems, Remove extensions, clear, remove item
Replace functionality in txteditor
Notifications missing files

- Change
TxtEditor auto size

- Fix
Two All in the Language comboBox
- Change

- Fix

- Bug