Big Update: Registry, Updates, Bug fixes
The last few pushes mainly revolved around the registry and adding a key to it. This works now so i will explain the basic changes.
I added a new section "Advanced" and added two buttons. One button ads a context menu item when you right click on a file, the other ads a context menu item when you right click on a folder. If you click on this item LaTexInclude will open with the file/files in the list. If you unchecked the "Start on context click", LaTexInclude will produce an output.tex file in your current position. This also means that the position of you LaTexInclude.exe has to stay in a fixed position else you won't be able to start the exe with the context menu item. Currently it's not possible to select multiple files and open it with LaTexInclude, however using the advanced settings you are able to select a folder and open everything in it in one app instance.

- Add
added new settings and a new seetings section
registry entry with a button click
auto generate output.tex when using the windows context menu if setting  is unchecked
automatic language selection for files by using the whitelist
missing file check
missing folder check
textbox check (only letters allowed in the box)

- Change
"All" will be selected on startup in the language box
normal textbox instead of masked one
extension "subarray"

- Bug fix
"All" disappearing when adding a new extension
maskedtextbox switching between instances ('.' and ',')
multiple instances appeared when selecting more than one file with the context menu